+++ android/src/main/java/com/reactnativestripesdk/customersheet/CustomerSheetFragment.kt
@@ -20,7 +20,6 @@ import com.stripe.android.customersheet.CustomerAdapter
 import com.stripe.android.customersheet.CustomerEphemeralKey
 import com.stripe.android.customersheet.CustomerSheet
 import com.stripe.android.customersheet.CustomerSheetResult
-import com.stripe.android.customersheet.ExperimentalCustomerSheetApi
 import com.stripe.android.customersheet.PaymentOptionSelection
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentsheet.*
@@ -29,7 +28,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 
-@OptIn(ExperimentalCustomerSheetApi::class, ExperimentalAllowsRemovalOfLastSavedPaymentMethodApi::class)
+@OptIn(ExperimentalAllowsRemovalOfLastSavedPaymentMethodApi::class)
 class CustomerSheetFragment : Fragment() {
   private var customerSheet: CustomerSheet? = null
   internal var customerAdapter: ReactNativeCustomerAdapter? = null
@@ -111,10 +110,11 @@ class CustomerSheetFragment : Fragment() {
 
     customerSheet = CustomerSheet.create(
       fragment = this,
-      configuration = configuration.build(),
       customerAdapter = customerAdapter,
       callback = ::handleResult
     )
+
+    customerSheet?.configure(configuration.build())
 
     initPromise.resolve(WritableNativeMap())
   }